functions/step_3/text_to_model/requirementss.txt
common @ git+https://github.com/it-minds/itm_leads-finder-scripts.git@main#subdirectory=internal
langchain-core==0.2.3
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-servicebus
pydantic==2.7.2